+++ src/fable-library/System.Collections.Generic.fs
@@ -113,7 +113,12 @@ type Queue<'T> private (initialContents, initialCount) =
     let mutable contents: 'T array = initialContents
     let mutable count = initialCount
     let mutable head = 0
-    let mutable tail = initialCount
+
+    let mutable tail =
+        if initialCount = contents.Length then
+            0
+        else
+            initialCount
 
     let size () = contents.Length
 
@@ -129,8 +134,13 @@ type Queue<'T> private (initialContents, initialCount) =
             Array.blit contents 0 newBuffer (size () - head) tail
 
         head <- 0
-        tail <- count
         contents <- newBuffer
+
+        tail <-
+            if count = size () then
+                0
+            else
+                count
 
     let toSeq () =
         seq {