+++ src/Fable.Cli/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [GH-3668](https://github.com/fable-compiler/Fable/pull/3668) Normalize fable-library argument (by @nojaf)
 
+#### Javascript
+
+* Fixed 'System.Collections.Generic.Queue' bug (by @PierreYvesR)
+
 #### Python
 
 * Fixed nested type with custom hashcode (by @dbrattli)
@@ -23,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed unary negation for signed integer MinValue (by @ncave)
 * Fixed excluding signature files from imports (by @ncave)
 * Fixed generic try_catch closure trait (by @ncave)
+* Fixed 'System.Collections.Generic.Queue' bug (by @PierreYvesR)
 
 ## 4.9.0 - 2023-12-14